% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytheme.R
\name{shinytheme}
\alias{shinytheme}
\title{Return the URL for a Shiny theme}
\usage{
shinytheme(theme = NULL)
}
\arguments{
\item{theme}{Name of a theme.}
}
\description{
The result of this function should be used as the \code{theme} argument for
\code{\link[shiny]{bootstrapPage}}, \code{\link[shiny]{fluidPage}},
\code{\link[shiny]{navbarPage}}, or \code{\link[shiny]{fixedPage}}.
}
\examples{
\dontrun{
shinyApp(
  ui = fluidPage(theme = shinytheme("united"),
    ...
  ),
  server = function(input, output) { }
)
}
}
\seealso{
The main \link{shinythemes} page for information about available
  themes and more detailed examples.
}
